requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
alembic>=1.13.1
asyncpg>=0.30.0
pydantic-settings>=2.2.1
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
httpx>=0.27.0
